Make "inline" unconditional
