+++ Cargo.toml
@@ -16,7 +16,6 @@ rust-version = "1.60"
 
 [features]
 default = ["std"]
-inline = []
 std = ["num-integer/std", "num-traits/std"]
 arbitrary = ["dep:arbitrary"]
 quickcheck = ["dep:quickcheck"]
@@ -24,7 +23,7 @@ rand = ["dep:rand"]
 serde = ["dep:serde"]
 
 [package.metadata.docs.rs]
-features = ["std", "inline", "serde", "rand", "quickcheck", "arbitrary"]
+features = ["std", "serde", "rand", "quickcheck", "arbitrary"]
 rustdoc-args = ["--cfg", "docsrs"]
 
 [[bench]]
